chore: Split header into components, Fixes #845
