+++ src/components/layout/header/development-warning.module.scss
@@ -1,4 +1,4 @@
-@import '../../scss/colors.module.scss';
+@import '~scss/colors.module.scss';
 
 .development-warning {
   color: white;
