+++ src/components/layout/header/navigation.js
@@ -2,7 +2,7 @@
 import React from 'react'
 import { Link } from 'gatsby'
 import { Menu, MenuList, MenuButton, MenuLink } from '@reach/menu-button'
-import headerNavigationStyles from './header-navigation.module.scss'
+import headerNavigationStyles from './navigation.module.scss'
 
 const MenuCaret = ({
   children,
