+++ src/components/layout/header/partnership-banner.js
