+++ src/components/layout/header/search-autocomplete.module.scss
@@ -1,7 +1,7 @@
-@import '../../scss/colors.module.scss';
+@import '~scss/colors.module.scss';
 
 .popover {
-  background: #FFFFFF;
+  background: #ffffff;
   border: 1px solid $color-slate-600;
   font-size: 1rem;
   position: absolute;
@@ -14,7 +14,7 @@
   opacity: 1;
   transform: translateY(0.4rem);
   color: $text;
-  
+
   [data-user-value] {
     font-weight: bold;
   }
@@ -28,7 +28,7 @@
     &:hover,
     &:focus,
     &[aria-selected='true'] {
-      background: transparentize($color-blueberry-200, .77);
+      background: transparentize($color-blueberry-200, 0.77);
     }
   }
 }
